errores refraccion

+++ interface/forms/eye_mag/save.php
@@ -1190,9 +1190,20 @@ if ($requestMode == "new") {
         }
     }
     //now save any Wear RXs (1-4) entered.
-
-    $rx_number = '1';
-    if ($_POST['W_1'] == '1') {
+    //Guard this block to avoid deleting data when Wear fields are not posted (e.g. disabled/read-only form).
+    $has_wearing_payload = isset($_POST['W_1']) || isset($_POST['W_2']) || isset($_POST['W_3']) || isset($_POST['W_4']);
+    if ($has_wearing_payload) {
+        $rx_number = '1';
+        $comments_w1 = $_POST['COMMENTS_1'] ?? ($_POST['COMMENTS_W'] ?? '');
+        $odnearva_1 = $_POST['ODNEARVA_1'] ?? ($_POST['NEARODVA_1'] ?? '');
+        $osnearva_1 = $_POST['OSNEARVA_1'] ?? ($_POST['NEAROSVA_1'] ?? '');
+        $odnearva_2 = $_POST['ODNEARVA_2'] ?? ($_POST['NEARODVA_2'] ?? '');
+        $osnearva_2 = $_POST['OSNEARVA_2'] ?? ($_POST['NEAROSVA_2'] ?? '');
+        $odnearva_3 = $_POST['ODNEARVA_3'] ?? ($_POST['NEARODVA_3'] ?? '');
+        $osnearva_3 = $_POST['OSNEARVA_3'] ?? ($_POST['NEAROSVA_3'] ?? '');
+        $odnearva_4 = $_POST['ODNEARVA_4'] ?? ($_POST['NEARODVA_4'] ?? '');
+        $osnearva_4 = $_POST['OSNEARVA_4'] ?? ($_POST['NEAROSVA_4'] ?? '');
+    if (isset($_POST['W_1']) && $_POST['W_1'] == '1') {
         $query = "REPLACE INTO `form_eye_mag_wearing` (`ENCOUNTER` ,`FORM_ID` ,`PID` ,`RX_NUMBER` ,`ODSPH` ,`ODCYL` ,`ODAXIS` ,
             `ODVA` ,`ODADD` ,`ODNEARVA` ,`OSSPH` ,`OSCYL` ,`OSAXIS` ,
             `OSVA` ,`OSADD` ,`OSNEARVA` ,`ODMIDADD` ,`OSMIDADD` ,
@@ -1205,19 +1216,19 @@ if ($requestMode == "new") {
             (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
             $LENS_TREATMENTS_1 = implode("|", (empty($_POST['LENS_TREATMENTS_1']) ? array() : $_POST['LENS_TREATMENTS_1']));
             sqlQuery($query, array($encounter, $form_id, $pid, $rx_number, $_POST['ODSPH_1'], $_POST['ODCYL_1'], $_POST['ODAXIS_1'],
-                $_POST['ODVA_1'], $_POST['ODADD_1'], $_POST['ODNEARVA_1'], $_POST['OSSPH_1'], $_POST['OSCYL_1'], $_POST['OSAXIS_1'],
-                $_POST['OSVA_1'], $_POST['OSADD_1'], $_POST['OSNEARVA_1'], $_POST['ODMIDADD_1'], $_POST['OSMIDADD_1'],
-                0 + $_POST['RX_TYPE_1'], $_POST['COMMENTS_1'],
+                $_POST['ODVA_1'], $_POST['ODADD_1'], $odnearva_1, $_POST['OSSPH_1'], $_POST['OSCYL_1'], $_POST['OSAXIS_1'],
+                $_POST['OSVA_1'], $_POST['OSADD_1'], $osnearva_1, $_POST['ODMIDADD_1'], $_POST['OSMIDADD_1'],
+                0 + $_POST['RX_TYPE_1'], $comments_w1,
                 $_POST['ODHPD_1'], $_POST['ODHBASE_1'], $_POST['ODVPD_1'], $_POST['ODVBASE_1'], $_POST['ODSLABOFF_1'], $_POST['ODVERTEXDIST_1'],
                 $_POST['OSHPD_1'], $_POST['OSHBASE_1'], $_POST['OSVPD_1'], $_POST['OSVBASE_1'], $_POST['OSSLABOFF_1'], $_POST['OSVERTEXDIST_1'],
                 $_POST['ODMPDD_1'], $_POST['ODMPDN_1'], $_POST['OSMPDD_1'], $_POST['OSMPDN_1'], $_POST['BPDD_1'], $_POST['BPDN_1'], $_POST['LENS_MATERIAL_1'],
                 $LENS_TREATMENTS_1));
             $rx_number++;
-    } else {
+    } elseif (isset($_POST['W_1'])) {
         $query = "DELETE FROM form_eye_mag_wearing where ENCOUNTER=? and PID=? and FORM_ID=? and RX_NUMBER=?";
         sqlQuery($query, array($encounter, $pid, $form_id, '1'));
     }
-    if ($_POST['W_2'] == '1') {
+    if (isset($_POST['W_2']) && $_POST['W_2'] == '1') {
         //store W_2
         $query = "REPLACE INTO `form_eye_mag_wearing` (`ENCOUNTER` ,`FORM_ID` ,`PID` ,`RX_NUMBER` ,`ODSPH` ,`ODCYL` ,`ODAXIS` ,
         `ODVA` ,`ODADD` ,`ODNEARVA` ,`OSSPH` ,`OSCYL` ,`OSAXIS` ,
@@ -1228,21 +1239,21 @@ if ($requestMode == "new") {
         `ODMPDD`,`ODMPDN`,`OSMPDD`,`OSMPDN`,`BPDD`,`BPDN`,`LENS_MATERIAL`,
         `LENS_TREATMENTS`
         ) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-        $LENS_TREATMENTS_2 = implode("|", $_POST['LENS_TREATMENTS_2']);
+        $LENS_TREATMENTS_2 = implode("|", (empty($_POST['LENS_TREATMENTS_2']) ? array() : $_POST['LENS_TREATMENTS_2']));
         sqlQuery($query, array($encounter, $form_id, $pid, $rx_number, $_POST['ODSPH_2'], $_POST['ODCYL_2'], $_POST['ODAXIS_2'],
-            $_POST['ODVA_2'], $_POST['ODADD_2'], $_POST['ODNEARVA_2'], $_POST['OSSPH_2'], $_POST['OSCYL_2'], $_POST['OSAXIS_2'],
-            $_POST['OSVA_2'], $_POST['OSADD_2'], $_POST['OSNEARVA_2'], $_POST['ODMIDADD_2'], $_POST['OSMIDADD_2'],
+            $_POST['ODVA_2'], $_POST['ODADD_2'], $odnearva_2, $_POST['OSSPH_2'], $_POST['OSCYL_2'], $_POST['OSAXIS_2'],
+            $_POST['OSVA_2'], $_POST['OSADD_2'], $osnearva_2, $_POST['ODMIDADD_2'], $_POST['OSMIDADD_2'],
             0 + $_POST['RX_TYPE_2'], $_POST['COMMENTS_2'],
             $_POST['ODHPD_2'], $_POST['ODHBASE_2'], $_POST['ODVPD_2'], $_POST['ODVBASE_2'], $_POST['ODSLABOFF_2'], $_POST['ODVERTEXDIST_2'],
             $_POST['OSHPD_2'], $_POST['OSHBASE_2'], $_POST['OSVPD_2'], $_POST['OSVBASE_2'], $_POST['OSSLABOFF_2'], $_POST['OSVERTEXDIST_2'],
             $_POST['ODMPDD_2'], $_POST['ODMPDN_2'], $_POST['OSMPDD_2'], $_POST['OSMPDN_2'], $_POST['BPDD_2'], $_POST['BPDN_2'], $_POST['LENS_MATERIAL_2'],
             $LENS_TREATMENTS_2));
         $rx_number++;
-    } else {
+    } elseif (isset($_POST['W_2'])) {
         $query = "DELETE FROM form_eye_mag_wearing where ENCOUNTER=? and PID=? and FORM_ID=? and RX_NUMBER=?";
         sqlQuery($query, array($encounter, $pid, $form_id, '2'));
     }
-    if ($_POST['W_3'] == '1') {
+    if (isset($_POST['W_3']) && $_POST['W_3'] == '1') {
         //store W_3
         $query = "REPLACE INTO `form_eye_mag_wearing` (`ENCOUNTER` ,`FORM_ID` ,`PID` ,`RX_NUMBER` ,`ODSPH` ,`ODCYL` ,`ODAXIS` ,
         `ODVA` ,`ODADD` ,`ODNEARVA` ,`OSSPH` ,`OSCYL` ,`OSAXIS` ,
@@ -1253,21 +1264,21 @@ if ($requestMode == "new") {
         `ODMPDD`,`ODMPDN`,`OSMPDD`,`OSMPDN`,`BPDD`,`BPDN`,`LENS_MATERIAL`,
         `LENS_TREATMENTS`
         ) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-        $LENS_TREATMENTS_3 = implode("|", $_POST['LENS_TREATMENTS_3']);
+        $LENS_TREATMENTS_3 = implode("|", (empty($_POST['LENS_TREATMENTS_3']) ? array() : $_POST['LENS_TREATMENTS_3']));
         sqlQuery($query, array($encounter, $form_id, $pid, $rx_number, $_POST['ODSPH_3'], $_POST['ODCYL_3'], $_POST['ODAXIS_3'],
-            $_POST['ODVA_3'], $_POST['ODADD_3'], $_POST['ODNEARVA_3'], $_POST['OSSPH_3'], $_POST['OSCYL_3'], $_POST['OSAXIS_3'],
-            $_POST['OSVA_3'], $_POST['OSADD_3'], $_POST['OSNEARVA_3'], $_POST['ODMIDADD_3'], $_POST['OSMIDADD_3'],
+            $_POST['ODVA_3'], $_POST['ODADD_3'], $odnearva_3, $_POST['OSSPH_3'], $_POST['OSCYL_3'], $_POST['OSAXIS_3'],
+            $_POST['OSVA_3'], $_POST['OSADD_3'], $osnearva_3, $_POST['ODMIDADD_3'], $_POST['OSMIDADD_3'],
             0 + $_POST['RX_TYPE_3'], $_POST['COMMENTS_3'],
             $_POST['ODHPD_3'], $_POST['ODHBASE_3'], $_POST['ODVPD_3'], $_POST['ODVBASE_3'], $_POST['ODSLABOFF_3'], $_POST['ODVERTEXDIST_3'],
             $_POST['OSHPD_3'], $_POST['OSHBASE_3'], $_POST['OSVPD_3'], $_POST['OSVBASE_3'], $_POST['OSSLABOFF_3'], $_POST['OSVERTEXDIST_3'],
             $_POST['ODMPDD_3'], $_POST['ODMPDN_3'], $_POST['OSMPDD_3'], $_POST['OSMPDN_3'], $_POST['BPDD_3'], $_POST['BPDN_3'], $_POST['LENS_MATERIAL_3'],
             $LENS_TREATMENTS_3));
         $rx_number++;
-    } else {
+    } elseif (isset($_POST['W_3'])) {
         $query = "DELETE FROM form_eye_mag_wearing where ENCOUNTER=? and PID=? and FORM_ID=? and RX_NUMBER=?";
         sqlQuery($query, array($encounter, $pid, $form_id, '3'));
     }
-    if ($_POST['W_4'] == '1') {
+    if (isset($_POST['W_4']) && $_POST['W_4'] == '1') {
         //store W_4
         $query = "REPLACE INTO `form_eye_mag_wearing` (`ENCOUNTER` ,`FORM_ID` ,`PID` ,`RX_NUMBER` ,`ODSPH` ,`ODCYL` ,`ODAXIS` ,
         `ODVA` ,`ODADD` ,`ODNEARVA` ,`OSSPH` ,`OSCYL` ,`OSAXIS` ,
@@ -1278,17 +1289,17 @@ if ($requestMode == "new") {
         `ODMPDD`,`ODMPDN`,`OSMPDD`,`OSMPDN`,`BPDD`,`BPDN`,`LENS_MATERIAL`,
         `LENS_TREATMENTS`
         ) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-        $LENS_TREATMENTS_4 = implode("|", $_POST['LENS_TREATMENTS_4']);
+        $LENS_TREATMENTS_4 = implode("|", (empty($_POST['LENS_TREATMENTS_4']) ? array() : $_POST['LENS_TREATMENTS_4']));
         sqlQuery($query, array($encounter, $form_id, $pid, $rx_number, $_POST['ODSPH_4'], $_POST['ODCYL_4'], $_POST['ODAXIS_4'],
-            $_POST['ODVA_4'], $_POST['ODADD_4'], $_POST['ODNEARVA_4'], $_POST['OSSPH_4'], $_POST['OSCYL_4'], $_POST['OSAXIS_4'],
-            $_POST['OSVA_4'], $_POST['OSADD_4'], $_POST['OSNEARVA_4'], $_POST['ODMIDADD_4'], $_POST['OSMIDADD_4'],
+            $_POST['ODVA_4'], $_POST['ODADD_4'], $odnearva_4, $_POST['OSSPH_4'], $_POST['OSCYL_4'], $_POST['OSAXIS_4'],
+            $_POST['OSVA_4'], $_POST['OSADD_4'], $osnearva_4, $_POST['ODMIDADD_4'], $_POST['OSMIDADD_4'],
             0 + $_POST['RX_TYPE_4'], $_POST['COMMENTS_4'],
             $_POST['ODHPD_4'], $_POST['ODHBASE_4'], $_POST['ODVPD_4'], $_POST['ODVBASE_4'], $_POST['ODSLABOFF_4'], $_POST['ODVERTEXDIST_4'],
             $_POST['OSHPD_4'], $_POST['OSHBASE_4'], $_POST['OSVPD_4'], $_POST['OSVBASE_4'], $_POST['OSSLABOFF_4'], $_POST['OSVERTEXDIST_4'],
             $_POST['ODMPDD_4'], $_POST['ODMPDN_4'], $_POST['OSMPDD_4'], $_POST['OSMPDN_4'], $_POST['BPDD_4'], $_POST['BPDN_4'], $_POST['LENS_MATERIAL_4'],
             $LENS_TREATMENTS_4));
         $rx_number++;
-    } else {
+    } elseif (isset($_POST['W_4'])) {
         $query = "DELETE FROM form_eye_mag_wearing where ENCOUNTER=? and PID=? and FORM_ID=? and RX_NUMBER=?";
         sqlQuery($query, array($encounter, $pid, $form_id, '4'));
     }
@@ -1296,6 +1307,7 @@ if ($requestMode == "new") {
     for ($i = $rx_number; $i < 5; $i++) {
         $query = "DELETE FROM form_eye_mag_wearing where ENCOUNTER=? and PID=? and FORM_ID=? and RX_NUMBER=?";
         sqlQuery($query, array($encounter, $pid, $form_id, $i));
+    }
     }
     //now return the obj
     $send['IMPPLAN_items'] = build_IMPPLAN_items($pid, $form_id);
